Remove required dependency on OpenAPI - use Javalin OpenAPI annotations.
Previously, I could not find a way to apply the built-in Javalin OpenAPI annotations to controller methods because they were not `Handler`s. In order for Javalin to extract OpenAPI annotations from a method or lambda, it must have the signature `(ctx: Context) -> void`; however controller methods can accept any arbitrary set of arguments. I discovered a helper method provided by Javalin under the `OpenApiBuilder` class, called `moveDocumentationFromAnnotationToHandler`, that allows OpenAPI on an arbitrary method to be associated with a handler, so now I simply call this method after creating the handler in the code generation piece. This means I can retire all of the OpenAPI annotations and code generation I wrote previously. This also means that the dependencies required for OpenAPI also go away.