Add WebSocket initial support.
WebSockets work very differently than HTTP requests. WebSockets are concerned with connectivity, so there are multiple handlers on a single route for detecting connections opening and closing, and client errors.  There's less concern about having complex model-binding and action results, although this might be a future effort. For now, I just want to have controller classes that can be instantiated via D.I. As part of this task, I realized that I cannot pass the context as an argument to the D.I. setup logic. I made the hard decision to hard-code all Javalin-specific binding rather than use the D.I. framework to create these objects. This means there's no need to inherit from a `ControllerComponent` or force users to add a `JavalinControllerModule` to their D.I. setup. In fact, after this change, I feel one step closer to perhaps freeing users from needing to use one particular D.I. framework.

In the future, I feel like I should focus on model binding of some sort, but I am not sure how this fits in with the other handlers. It could be as simple as creating an abstract `WsController` class that implements `onMessage` and you pass it a lambda taking the desired arguments... and it binds them on the fly. We'll see!