Better handling of unbound primitive parameters.
Previously, if a primitive parameter could not be bound to a value source, a null was returned. However, this immediately led to a NPE when trying to cast the value and pass it to the controller method. Since this NPE happens inside generated code, it would be almost impossible to diagnose the underlying issue - that the value could not be bound. For now, the simple solution is to simply initial the primitive to the default value (0 initialized). In the future, we could throw a 400 or 500 error; however, the recommended best practice is to always use boxed primitives so null can be returned, detected, and handled. This approach doesn't dictate how the error should be handled but allows for checking for 0, false, etc.