Replicate cookie primitive unit tests for header params.
I wanted to see how hard it would be to take the unit tests I wrote for cookies and just make them work with headers. It was still a bit of work but it did uncover some other strange issues that are included in previous commits. I also added some tests for handle unbound primitive types to test out changes made in previous commits.