Create ControllerRegistry interface.
I used to call the generated class ControllerRegistry. However, it's useful to define an interface so you can write most of the code before you compile the first time. The concrete JavalinControllerRegistry is created at compile time. When the compilation fails now, the only error is the concrete type not being define. In the future, this might also allow us to create subclasses that work for specific versions of Javalin.