(WIP) First working example calling a custom conversion method.
This commit includes the first passing unit test calling a custom conversion method. This commit also includes duplicate route checking for WebSocket endpoints and duplicate converter names.

The next step it to add support for custom converters for WebSocket endpoints, then add support for custom converters on model fields/setters. That should be the end of it.