Capture exceptions parsing JSON.
I want to capture exceptions when parsing JSON. Otherwise, the error is being thrown within generated code which is harder to diagnose. It's much easier to receive a null object and debug the cause. Another thought would be to allow this exception to be thrown and expect the developer to use a @After annotation to inspect the exception. I might want to add logging to report the failure.