(WIP) Avoid generating code for retrieving injector, unless used somewhere.
While creating a Dagger container is basically "free", seeing a potentially unused variable getting generated is kind of gross. It kind of leads to a pattern where every "thing" that can be generated returns a "result" object. I could see this result object becoming something more complex in the future. I need to the ability to pass around variable names to avoid hard-coding everywhere, and perhaps a better approach is to create classes taking input and providing methods for getting results.