More unit testing and bug fixes. Initial dynamic DI experiments.
I was missing a test to make sure model parameters requiring DI to be created existed... and of course there was a bug. These model builders are put in helper methods, which can call each other recursively, and I needed to pass the DI container down the levels.

I had a theory that I was not handling controller action methods returning void. I handled void in a couple places, but not universally. Primarily, I needed to handle returning `CompletableFuture<Void>`. I had a lack of tests for different return types, in general, so I beefed up my coverage there.

I found I had a bug related to passing in `WsRequest` and `WsResponse` to Websocket message methods.

The next step to supporting runtime DI libraries, like Guice, is going to be introducing my own interface for DI.