Request class name consistency and helper methods.
I wanted the method names for working with path, query string, header, cookie and form data values to be consistent. For v5, I will retire the old names entirely but am just deprecating them for now. I updated the code generation to use the name method names.

I also wanted to provide convenience methods for getting any of the value sources as `List<String>` even if that source doesn't support multiple values per key. In that case, it returns a singleton list or an empty list.