Support @DefaultValue for collection action method parameters
For collection types, the @DefaultValue is returned as a singleton containing the specified default value. I had a bug in the scalar implementation where, instead of passing the `defaultValue` parameter down the call stack, I was passing the actual default value. This means in the implementation, a cached method would have been created referring to whatever the first default value was.