Switch to Jakarta. Use new OpenAPI dependencies (WIP). Distinct methods for handlers.
There's a new OpenAPI dependency for Javalin. It looks like it only works for types, methods, and fields, so I need to create actual methods for each handler. I will need to copy the `@OpenApi` annotation from the controller method to the new method. I will do this in a future commit, since it will be hard, probably. I will need to update the README too.

There's several methods that got renamed and enums introduced, so I dealt with that in this commit. Everything should compile now, at least.

The javax package was replaced with jakarta, for the most part. I started replacing the annotations here.