Fix inexplicable preprocessing error related to common-lang3.
As much as I hate replicating the code in apache commons, I was getting a very strange error related to MutableInt#getAndIncrement not existing when Intellij was trying to build the project. I don't mind removing a dependency that might lead to conflicts in other projects and the implementations are seemingly very straight-forward.